refactor: AnimatedTitle, MemorialStone 관심사 분리
