+++ client/src/GalleryPage/mapObjects/MemorialStone.tsx
@@ -1,12 +1,79 @@
-import { Billboard, Text } from "@react-three/drei";
-import { useFrame, useThree } from "@react-three/fiber";
 import { useEffect, useMemo, useRef, useState } from "react";
-import { Vector3 } from "three";
+import { Text } from "@react-three/drei";
+import { useFrame, useThree } from "@react-three/fiber";
+import { Mesh, Vector3 } from "three";
+
 import { IGalleryPageSubTitle } from "../../@types/gallery";
 import MapoFont from "../../assets/MapoFlowerIsland.otf";
+
+interface MemorialStonesProps {
+  subtitles: IGalleryPageSubTitle[];
+  position: number[];
+}
+
 interface MemorialStoneProps {
   subTitle: IGalleryPageSubTitle;
   position: number[];
+}
+
+interface IStoneInfo {
+  subtitle: IGalleryPageSubTitle;
+  stonePosition: number[];
+}
+
+function calculateMemorialStonePosition(subtitles: IGalleryPageSubTitle[]) {
+  const enalblePositions = [
+    [0, 0],
+    [1, 1],
+    [-1, -1],
+    [2, 0],
+    [-2, 0],
+    [3, 1],
+    [-3, -1],
+  ];
+  let cursor = 0;
+  const stoneInfoList: IStoneInfo[] = [];
+  const h1List = subtitles.filter((subTitle: IGalleryPageSubTitle) => subTitle.type === "h1");
+  const h2List = subtitles.filter((subTitle: IGalleryPageSubTitle) => subTitle.type === "h2");
+  const h3List = subtitles.filter((subTitle: IGalleryPageSubTitle) => subTitle.type === "h3");
+  while (cursor < enalblePositions.length) {
+    if (h1List.length > 0) {
+      const subtitle = h1List.pop();
+      if (subtitle) {
+        stoneInfoList.push({
+          subtitle,
+          stonePosition: enalblePositions[cursor],
+        });
+        cursor++;
+        continue;
+      }
+    } else if (h2List.length > 0) {
+      const subtitle = h2List.pop();
+      if (subtitle) {
+        stoneInfoList.push({
+          subtitle,
+          stonePosition: enalblePositions[cursor],
+        });
+        cursor++;
+        continue;
+      }
+      cursor++;
+      continue;
+    } else if (h3List.length > 0) {
+      const subtitle = h3List.pop();
+      if (subtitle) {
+        stoneInfoList.push({
+          subtitle,
+          stonePosition: enalblePositions[cursor],
+        });
+        cursor++;
+        continue;
+      }
+    } else {
+      break;
+    }
+  }
+  return stoneInfoList;
 }
 
 function textPreProcessing(text: string) {
@@ -22,7 +89,7 @@ function textPreProcessing(text: string) {
   return { visibleLetters, invisibleLetters };
 }
 
-export default function MemorialStone({ subTitle, position }: MemorialStoneProps) {
+function MemorialStone({ subTitle, position }: MemorialStoneProps) {
   const { text, type } = subTitle;
   const [pText, setPText] = useState("");
   const { camera } = useThree();
@@ -33,7 +100,7 @@ export default function MemorialStone({ subTitle, position }: MemorialStoneProps
     return { visibleLetters, invisibleLetters };
   }, []);
 
-  const subtitleMeshRef = useRef<THREE.Mesh>(null);
+  const subtitleMeshRef = useRef<Mesh>(null);
   const subtitleRef = useRef<any>();
 
   useFrame(() => {
@@ -86,6 +153,27 @@ export default function MemorialStone({ subTitle, position }: MemorialStoneProps
           {pText}
         </Text>
       </mesh>
+    </>
+  );
+}
+
+export default function MemorialStones({ subtitles, position }: MemorialStonesProps) {
+  const stoneInfoList = useMemo(() => {
+    return calculateMemorialStonePosition(subtitles);
+  }, []);
+  return (
+    <>
+      {stoneInfoList.map((stoneInfo, i) => {
+        const { subtitle, stonePosition } = stoneInfo;
+        const key = `${subtitle}+${i}`;
+        return (
+          <MemorialStone
+            subTitle={subtitle}
+            position={position.map((e: number, i: number) => e + stonePosition[i])}
+            key={key}
+          />
+        );
+      })}
     </>
   );
 }
